[xs_modules] Added IR sensing to autodock
Added infrared sensing capabiliites to the autodocking method. The
robot will now check to see if it is near the dock by checking the IR
sensor information. If it can not see the IR signals, it will rotate
around 180deg (depending on how fast it is set to turn) and check
again.